[Fix] remove protocol
